# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HACK",
        "version": "1.0"
    },
    "paths": {
...
        "/api/logs/{id_action}": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Retrieve logs by action ID",
                "operationId": "39ae429ff74169955384d24a6282721c",
                "parameters": [
                    {
                        "name": "id_action",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "action": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "user_name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No logs found for this action ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No logs found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-logs/{id_user}": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Retrieve all logs for a specific user",
                "operationId": "91875cbc7789bad6ad43f50a33381638",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "action": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "user_name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No logs found for this user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No logs found for this user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/generate-password": {
            "get": {
                "tags": [
                    "Password"
                ],
                "summary": "Generate a secure password",
                "description": "Generate a secure password.",
                "operationId": "fb8d50b2dc70818bf841d978152f42b0",
                "parameters": [
                    {
                        "name": "length",
                        "in": "query",
                        "description": "Desired password length (default is 12)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secure password generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "password": {
                                            "type": "string",
                                            "example": "aZ3$w8@kF"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid length provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Password length must be at least 8"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/check-common-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Check if a password is common and insecure",
                "description": "Verifies if the provided password is in a list of the most common passwords, making it insecure.",
                "operationId": "2b6f2b040644a2107daf20d9a2820793",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password is not in the list of common passwords",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This password is not in the list of common passwords."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password is too common and insecure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This password is too common and insecure."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving the password list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unable to retrieve password list."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}